Expose server spec so it can be used by InfernoOperationsM implementors
